+++ turbo/apps/host-worker/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [1.5.7](https://github.com/vm0-ai/okou/compare/host-worker-v1.5.6...host-worker-v1.5.7) (2026-09-20)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @okouai/api-contracts bumped to 1.477.0
+
 ## [1.5.6](https://github.com/vm0-ai/okou/compare/host-worker-v1.5.5...host-worker-v1.5.6) (2026-09-20)